+++ crates/debugger_ui/src/session/running/stack_frame_list.rs
